# doc/openapi.yaml
openapi: 3.0.0
info:
  title: 'Miniblog Api'
  contact:
    email: rayjlim1@gmail.com
  version: '0.1'
paths:
  '/api/posts/{id}':
    put:
      description: 'Update Entry'
      operationId: 'CUDHandler::updateEntry'
      parameters:
        -
          name: passengers
          in: query
          description: 'passenger information'
          required: true
          schema:
            type: array
            items:
              properties:
                firstName:
                  description: 'The product name'
                lastName: {  }
                middleName: {  }
              type: object
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResults'
    post:
      description: 'Create New Entry'
      operationId: 'CUDHandler::addEntry'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResults'
    delete:
      description: 'Remove Entry'
      operationId: 'CUDHandler::deleteEntry'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResults'
  /api/posts/:
    get:
      operationId: 'EntryHandler::listItemsApi'
      responses:
        '200':
          description: 'Retrieve entries limit 50'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SmsEntrie'
  /api/sameDayEntries:
    get:
      description: 'Retrieve entries on same day of year'
      operationId: 'EntryHandler::sameDayEntries'
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResults'
        '404':
          description: 'Could Not Find Resource'
  /api/yearMonth:
    get:
      description: 'Retrieve entries limit 50'
      operationId: 'EntryHandler::yearMonthsApi'
      requestBody:
        description: 'Client side search object'
        required: true
        content:
          application/json: {  }
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResults'
        '404':
          description: 'Could Not Find Resource'
components:
  schemas:
    SearchResults:
      type: array
      items:
        $ref: '#/components/schemas/SmsEntrie'
    SmsEntrie:
      properties:
        id:
          description: 'The entry id'
          type: integer
        content:
          description: 'The entry content'
          type: string
        date:
          description: 'The entry date'
          type: string
          format: date
        userId:
          description: 'The entry creator id'
          type: integer
      type: object